+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Major Bugfix: Issue in tracker where annotations were not read in the requested resolution causing scores of zero.
 * Fixed issue where DiceFocalLoss was not passed class weights correctly.
 * We no longer require a hacked pycold
+* Demo region model timestamps no longer depend on the locale.
 
 
 ### Changed